Augment number of epochs for single network sscsd

+++ Evaluation/mcsd_evaluations.py
@@ -35,8 +35,8 @@ def main():
             if is_std_method:
                 std_mcsd.main(data_path, bvals, bvecs, mask_path, order, save_to)
             else:
-                multi_network_sscsd.main(data_path, bvals, bvecs, mask_path, order, False, save_to)
-                # single_network_sscsd.main(data_path, bvals, bvecs, mask_path, order, False, save_to)
+                # multi_network_sscsd.main(data_path, bvals, bvecs, mask_path, order, False, save_to)
+                single_network_sscsd.main(data_path, bvals, bvecs, mask_path, order, False, save_to)
             dipy_peak_extraction.peak_extraction(
                 save_to + '/odfs.nii.gz',
                 'sphere724.txt',
@@ -46,7 +46,8 @@ def main():
                 max_peak_number=3
             )
             os.remove(save_to + '/odfs.nii.gz')
-        os.remove(save_to + '/weights.pth')
+        if not is_std_method:
+            os.remove(save_to + '/weights.pth')
 
 
 if __name__ == '__main__':
